feat: implement MessageInput component with dynamic height and send functionality
